+++ commfuncs.h
@@ -28,7 +28,7 @@
 
 using namespace std;
 
-#define VERSION "v0.992"
+#define VERSION "v0.993"
 
 #define UNKNOWN 0