+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 All change logs will be documented here.<br/>
-   # [0.991] 2022-11-02
+   # [0.993] 2022-11-02
+   Enhancement: getpart will return empty string if the given index is out of range. <br/>
+   Bux fix: when function does not work properly if the condition filter is not const. <br/>
+   # [0.992] 2022-11-02
    Enhancement: Compare two ANY date type data as STRING. <br/>
    # [0.991] 2022-11-01
    New feature: new query option: report|r SelectionIndex1:AggregationOp1[,SelectionIndex2:AggregationOp2] -- Generate a summary of specified selections. <br/>